fix(ci): line-limit check — split oversized GitHub client tests
